Fix update routines (check if column/tables exist)

+++ database.sql
@@ -1,6 +1,6 @@
 -- ------------------------------------------
 -- Friendica 2020.09-dev (Red Hot Poker)
--- DB_UPDATE_VERSION 1368
+-- DB_UPDATE_VERSION 1367
 -- ------------------------------------------
 
 
@@ -589,7 +589,7 @@ CREATE TABLE IF NOT EXISTS `hook` (
 -- TABLE host
 --
 CREATE TABLE IF NOT EXISTS `hosts` (
-	`id` tinyint unsigned NOT NULL auto_increment COMMENT 'sequential ID',
+	`id` mediumint unsigned NOT NULL auto_increment COMMENT 'sequential ID',
 	`name` varchar(128) NOT NULL DEFAULT '' COMMENT 'The hostname',
 	 PRIMARY KEY(`id`),
 	 UNIQUE INDEX `name` (`name`)
@@ -799,7 +799,7 @@ CREATE TABLE IF NOT EXISTS `locks` (
 	`id` int unsigned NOT NULL auto_increment COMMENT 'sequential ID',
 	`name` varchar(128) NOT NULL DEFAULT '' COMMENT '',
 	`locked` boolean NOT NULL DEFAULT '0' COMMENT '',
-	`host-id` tinyint unsigned NOT NULL DEFAULT 0 COMMENT 'Host id',
+	`host-id` mediumint unsigned NOT NULL DEFAULT 0 COMMENT 'Host id',
 	`pid` int unsigned NOT NULL DEFAULT 0 COMMENT 'The process id of the worker',
 	`expires` datetime NOT NULL DEFAULT '0001-01-01 00:00:00' COMMENT 'datetime of cache expiration',
 	 PRIMARY KEY(`id`),
